20230503 고객센터 1.권한설정제외하고 다 완성
여숙맛
1.리뷰멤버 닉네임 으로 보이게한것
2.로그인하면 로그인한 계정의 닉네임으로 리뷰 작성 보이게

+++ s20230404/src/main/java/com/travelAlone/s20230404/model/mh/Notice.java
@@ -12,7 +12,7 @@ import lombok.ToString;
 public class Notice {
 	
 	private int 				g_notice_id; 
-	private int 				member_id;
+	private long 				member_id;
 	private String 				g_common_csboard;
 	private String 				g_notice_title;
 	private String 				g_notice_content;
